+++ bundles/org.eclipse.swt/Eclipse SWT PI/cocoa/org/eclipse/swt/internal/cocoa/SWTPrintPanelDelegate.java
@@ -16,7 +16,7 @@ public SWTPrintPanelDelegate() {
 	super(0);
 }
 	
-public SWTPrintPanelDelegate(int id) {
+public SWTPrintPanelDelegate(long /*int*/ id) {
 	super(id);
 }
